+++ tests/CMakeLists.txt
@@ -57,6 +57,9 @@ target_link_libraries(test_lut PRIVATE loki-b)
 add_executable(test_parser test_parser.cpp)
 target_link_libraries(test_parser PRIVATE loki-b)
 
+add_executable(loki_test_range test_range.cpp)
+target_link_libraries(loki_test_range PRIVATE loki-b)
+
 add_executable(test_reaction_format test_reaction_format.cpp)
 target_link_libraries(test_reaction_format PRIVATE loki-b)
 
@@ -96,6 +99,7 @@ target_add_test(test_grid)
 target_add_test(test_linalg)
 target_add_test(test_lut)
 target_add_test(test_parser)
+target_add_test(loki_test_range)
 target_add_test(test_reaction_format)
 target_add_test(test_statebase)
 target_add_test(test_trans_coefs)
